requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
bcrypt==4.1.2
cryptography==41.0.7
paramiko==3.0.0
python-dotenv==1.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
psycopg2-binary==2.9.5
email-validator==2.0.0.post2
gunicorn==20.1.0
python-socketio==5.11.1
eventlet==0.35.2